+++ include/geometry2D/primitives/rect.inl
@@ -461,7 +461,7 @@ rect<T> rect<T>::united(const rect<T>& r) const
 template<class T>
 bool rect<T>::operator==(const rect& other) const
 {
-    for (uint d = 0; d < 2; d++)
+    for (size_t d = 0; d < 2; d++)
     {
         if (other._points[d] != _points[d])
         {
@@ -474,7 +474,7 @@ bool rect<T>::operator==(const rect& other) const
 template<class T>
 bool rect<T>::operator!=(const rect& other) const
 {
-    for (uint d = 0; d < 2; d++)
+    for (size_t d = 0; d < 2; d++)
     {
         if (other._points[d] != _points[d])
         {
